Move Participant and ParticipantLocation down into client crate
